backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-dotenv>=1.0.1
requests>=2.32.0
streamlit>=1.36.0
fireworks-ai>=0.19.0
pandas>=2.2.0 